# Conflicts:
#	.wiki/log.md
#	backend/src/main/java/com/xiyu/bid/analytics/controller/CompetitorAnalysisController.java
#	backend/src/main/java/com/xiyu/bid/analytics/dto/CompetitorAnalysisRequest.java
#	backend/src/main/java/com/xiyu/bid/analytics/dto/CompetitorAnalysisResponse.java
#	backend/src/main/java/com/xiyu/bid/analytics/dto/CompetitorAnalysisSeriesDTO.java
#	backend/src/main/java/com/xiyu/bid/analytics/model/CompetitorAnalysisRow.java
#	backend/src/main/java/com/xiyu/bid/analytics/service/CompetitorAnalysisAssemblerService.java
#	backend/src/main/java/com/xiyu/bid/analytics/service/CompetitorAnalysisComputationService.java
#	backend/src/main/java/com/xiyu/bid/analytics/service/CompetitorAnalysisQueryService.java
#	backend/src/main/java/com/xiyu/bid/analytics/service/CompetitorAnalysisService.java
#	backend/src/main/java/com/xiyu/bid/analytics/service/ProjectTypeAnalyticsQueryService.java
#	backend/src/test/resources/project-access-guard-baseline.txt
#	src/views/data-analysis/components/m0-kpi-cards/index.vue
#	src/views/data-analysis/components/m1-trend-analysis/DrillModal.vue
#	src/views/data-analysis/components/m1-trend-analysis/composables/useDrillDown.js
#	src/views/data-analysis/components/m1-trend-analysis/index.vue
#	src/views/data-analysis/components/m2-customer-type/index.vue
#	src/views/data-analysis/components/m3-project-type/index.vue
#	src/views/data-analysis/components/m4-competitor/chartRenderer.js
#	src/views/data-analysis/components/m4-competitor/composables/useCompetitorData.js
#	src/views/data-analysis/components/m4-competitor/index.vue
#	src/views/data-analysis/composables/useAnalyticsData.js
#	src/views/data-analysis/index.vue

+++ backend/src/main/java/com/xiyu/bid/analytics/dto/CompetitorAnalysisSeriesDTO.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class CompetitorAnalysisSeriesDTO {
     private String name;
     private String type;
-    private List<Integer> data;
+    private List<Double> data;
 }
